Member claim multi epoch multi feeHandler
