+++ test/deployment.js
@@ -12,6 +12,7 @@ const {
   ZERO_ADDRESS,
   NO_ZERO_ADDRESS,
   MAX_DELEGATION_FEE,
+  ETH_TOKEN_ADDRESS,
 } = require('./helper.js');
 
 let kyberPoolMaster;
@@ -29,21 +30,26 @@ contract('KyberPoolMaster deployment', async (accounts) => {
   describe('deployment', () => {
     it('should not allow to deploy a KyberPoolMaster zero address kyberDao', async () => {
       await expectRevert(
-        KyberPoolMaster.new(ZERO_ADDRESS, ZERO_ADDRESS, 0, 0),
+        KyberPoolMaster.new(
+          ZERO_ADDRESS,
+          0,
+          0,
+          [NO_ZERO_ADDRESS],
+          [ZERO_ADDRESS]
+        ),
         'ctor: kyberDao is missing'
-      );
-    });
-
-    it('should not allow to deploy a KyberPoolMaster zero address kyberFeeHandler', async () => {
-      await expectRevert(
-        KyberPoolMaster.new(NO_ZERO_ADDRESS, ZERO_ADDRESS, 0, 0),
-        'ctor: kyberFeeHandler is missing'
       );
     });
 
     it('should not allow to deploy a KyberPoolMaster with epochNotice lower than the minimum', async () => {
       await expectRevert(
-        KyberPoolMaster.new(NO_ZERO_ADDRESS, NO_ZERO_ADDRESS, 0, 0),
+        KyberPoolMaster.new(
+          NO_ZERO_ADDRESS,
+          0,
+          0,
+          [NO_ZERO_ADDRESS],
+          [ZERO_ADDRESS]
+        ),
         'ctor: Epoch Notice too low.'
       );
     });
@@ -52,11 +58,79 @@ contract('KyberPoolMaster deployment', async (accounts) => {
       await expectRevert(
         KyberPoolMaster.new(
           NO_ZERO_ADDRESS,
-          NO_ZERO_ADDRESS,
           1,
-          MAX_DELEGATION_FEE + 1
+          MAX_DELEGATION_FEE + 1,
+          [NO_ZERO_ADDRESS],
+          [ZERO_ADDRESS]
         ),
         'ctor: Delegation Fee greater than 100%'
+      );
+    });
+
+    it('should not allow to deploy a KyberPoolMaster without at least', async () => {
+      await expectRevert(
+        KyberPoolMaster.new(NO_ZERO_ADDRESS, 1, 1, [], [ZERO_ADDRESS]),
+        'ctor: at least one _kyberFeeHandlers require'
+      );
+    });
+
+    it('should not allow to deploy a KyberPoolMaster when feeHandlers and rewardToken are not even', async () => {
+      const kyberFeeHandler1 = await KyberFeeHandler.new(NO_ZERO_ADDRESS);
+      const kyberFeeHandler2 = await KyberFeeHandler.new(NO_ZERO_ADDRESS);
+      const ERC20RewartTokenA = await TestToken.new(
+        'Reward Token A',
+        'RTA',
+        18
+      );
+      await expectRevert(
+        KyberPoolMaster.new(
+          NO_ZERO_ADDRESS,
+          1,
+          1,
+          [kyberFeeHandler1.address, kyberFeeHandler2.address],
+          [ERC20RewartTokenA.address]
+        ),
+        'ctor: _kyberFeeHandlers and _rewardTokens uneven'
+      );
+      await expectRevert(
+        KyberPoolMaster.new(
+          NO_ZERO_ADDRESS,
+          1,
+          1,
+          [kyberFeeHandler1.address],
+          [ERC20RewartTokenA.address, ERC20RewartTokenA.address]
+        ),
+        'ctor: _kyberFeeHandlers and _rewardTokens uneven'
+      );
+    });
+
+    it('should not allow to deploy a KyberPoolMaster with ZERO_ADDRESS feeHandler', async () => {
+      await expectRevert.unspecified(
+        KyberPoolMaster.new(
+          NO_ZERO_ADDRESS,
+          1,
+          1,
+          [ZERO_ADDRESS],
+          [NO_ZERO_ADDRESS]
+        )
+      );
+    });
+
+    it('should not allow to deploy a KyberPoolMaster with repeated feeHandler', async () => {
+      const kyberFeeHandler1 = await KyberFeeHandler.new(NO_ZERO_ADDRESS);
+      const ERC20RewartTokenA = await TestToken.new(
+        'Reward Token A',
+        'RTA',
+        18
+      );
+      await expectRevert.unspecified(
+        KyberPoolMaster.new(
+          NO_ZERO_ADDRESS,
+          1,
+          1,
+          [kyberFeeHandler1.address, kyberFeeHandler1.address],
+          [ERC20RewartTokenA.address, ERC20RewartTokenA.address]
+        )
       );
     });
 
@@ -65,11 +139,14 @@ contract('KyberPoolMaster deployment', async (accounts) => {
 
       const kyberFeeHandler = await KyberFeeHandler.new(kyberDao.address);
 
+      const ERC20RewartToken = await TestToken.new('Reward Token A', 'RTA', 18);
+
       kyberPoolMaster = await KyberPoolMaster.new(
         kyberDao.address,
-        kyberFeeHandler.address,
         2,
         1,
+        [kyberFeeHandler.address],
+        [ERC20RewartToken.address],
         {
           from: poolMasterOwner,
         }
@@ -84,8 +161,13 @@ contract('KyberPoolMaster deployment', async (accounts) => {
       const kyberStakingAddress = await kyberPoolMaster.kyberStaking();
       expect(kyberStakingAddress).to.equal(NO_ZERO_ADDRESS);
 
-      const kyberFeeHandlerAddress = await kyberPoolMaster.kyberFeeHandler();
+      const kyberFeeHandlerAddress = await kyberPoolMaster.feeHandlersList(0);
       expect(kyberFeeHandlerAddress).to.equal(kyberFeeHandler.address);
+
+      const rewardToken = await kyberPoolMaster.rewardTokenByFeeHandle(
+        kyberFeeHandler.address
+      );
+      expect(rewardToken).to.equal(ERC20RewartToken.address);
 
       const epochNotice = await kyberPoolMaster.epochNotice();
       expect(epochNotice.toString()).to.equal('2');
@@ -96,7 +178,58 @@ contract('KyberPoolMaster deployment', async (accounts) => {
       expect(delegationFee.applied).to.equal(true);
     });
 
-    it('poolMaster should be able to claim ERC20 tokens', async () => {
+    it('should set ETH as reward token when ZERO_ADDRESS is used', async () => {
+      const kyberDao = await KyberDao.new(NO_ZERO_ADDRESS, NO_ZERO_ADDRESS);
+
+      const kyberFeeHandler = await KyberFeeHandler.new(kyberDao.address);
+
+      kyberPoolMaster = await KyberPoolMaster.new(
+        kyberDao.address,
+        2,
+        1,
+        [kyberFeeHandler.address],
+        [ZERO_ADDRESS],
+        {
+          from: poolMasterOwner,
+        }
+      );
+
+      const rewardToken = await kyberPoolMaster.rewardTokenByFeeHandle(
+        kyberFeeHandler.address
+      );
+      expect(rewardToken).to.equal(ETH_TOKEN_ADDRESS);
+    });
+
+    it('poolMaster should not be able to claim a reward tokens', async () => {
+      const kyberDao = await KyberDao.new(NO_ZERO_ADDRESS, NO_ZERO_ADDRESS);
+      const kyberFeeHandler = await KyberFeeHandler.new(kyberDao.address);
+      const ERC20RewartToken = await TestToken.new('Reward Token A', 'RTA', 18);
+
+      kyberPoolMaster = await KyberPoolMaster.new(
+        kyberDao.address,
+        2,
+        1,
+        [kyberFeeHandler.address],
+        [ERC20RewartToken.address],
+        {
+          from: poolMasterOwner,
+        }
+      );
+
+      await ERC20RewartToken.transfer(
+        kyberPoolMaster.address,
+        mulPrecision(10)
+      );
+
+      await expectRevert(
+        kyberPoolMaster.claimErc20Tokens(ERC20RewartToken.address, mike, {
+          from: poolMasterOwner,
+        }),
+        'not allowed to claim rewardTokens'
+      );
+    });
+
+    it('poolMaster should be able to claim ERC20 non reward tokens', async () => {
       await erc20.transfer(kyberPoolMaster.address, mulPrecision(10), {
         from: mike,
       });
@@ -119,12 +252,14 @@ contract('KyberPoolMaster deployment', async (accounts) => {
       const kyberDao = await KyberDao.new(NO_ZERO_ADDRESS, NO_ZERO_ADDRESS);
 
       const kyberFeeHandler = await KyberFeeHandler.new(kyberDao.address);
+      const ERC20RewartToken = await TestToken.new('Reward Token A', 'RTA', 18);
 
       kyberPoolMaster = await KyberPoolMaster.new(
         kyberDao.address,
-        kyberFeeHandler.address,
         2,
         1,
+        [kyberFeeHandler.address],
+        [ERC20RewartToken.address],
         {
           from: poolMasterOwner,
         }
