+++ test/rewards.js
@@ -6,11 +6,16 @@ const KyberFeeHandlerWithRewardPerEposhSetter = artifacts.require(
 const KyberFeeHandlerWithClaimStakerReward = artifacts.require(
   'KyberFeeHandlerWithClaimStakerReward'
 );
+const KyberFeeHandlerWithClaimStakerRewardERC20 = artifacts.require(
+  'KyberFeeHandlerWithClaimStakerRewardERC20'
+);
 const KyberStakingWithgetStakerDataForEpoch = artifacts.require(
   'KyberStakingWithgetStakerDataForEpoch'
 );
 
 const PoolMasterNoFallbackMock = artifacts.require('PoolMasterNoFallbackMock');
+
+const TestToken = artifacts.require('Token.sol');
 
 const {expect} = require('chai');
 const {
@@ -23,11 +28,21 @@ const {
 
 const Reverter = require('./utils/reverter');
 
-const {NO_ZERO_ADDRESS} = require('./helper.js');
+const {
+  NO_ZERO_ADDRESS,
+  ZERO_ADDRESS,
+  precisionUnits,
+  ETH_TOKEN_ADDRESS,
+} = require('./helper.js');
 
 let kyberPoolMaster;
 let kyberDao;
-let kyberFeeHandler;
+let kyberFeeHandler1;
+let kyberFeeHandler2;
+let kyberFeeHandler3;
+let kyberFeeHandler4;
+let rewardTokenA;
+let rewardTokenB;
 let daoSetter;
 let poolMasterOwner;
 let bank;
@@ -42,7 +57,7 @@ contract('KyberPoolMaster claiming', async (accounts) => {
     await reverter.snapshot();
   });
 
-  describe('#getUnclaimedRewards', () => {
+  describe('#getUnclaimedRewards - getAllEpochWithUnclaimedRewards', () => {
     before('one time init', async () => {
       daoSetter = accounts[1];
       poolMasterOwner = accounts[2];
@@ -51,32 +66,54 @@ contract('KyberPoolMaster claiming', async (accounts) => {
 
       kyberDao = await KyberDao.new(NO_ZERO_ADDRESS, NO_ZERO_ADDRESS);
 
-      kyberFeeHandler = await KyberFeeHandlerWithRewardPerEposhSetter.new(
+      kyberFeeHandler1 = await KyberFeeHandlerWithRewardPerEposhSetter.new(
         kyberDao.address
       );
+      kyberFeeHandler2 = await KyberFeeHandlerWithRewardPerEposhSetter.new(
+        kyberDao.address
+      );
+      rewardTokenA = await TestToken.new('Reward Token A', 'RTA', 18);
+      kyberFeeHandler3 = await KyberFeeHandlerWithRewardPerEposhSetter.new(
+        kyberDao.address
+      );
+      rewardTokenB = await TestToken.new('Reward Token B', 'RTB', 18);
 
       kyberPoolMaster = await KyberPoolMaster.new(
         kyberDao.address,
-        kyberFeeHandler.address,
         2,
         1,
+        [
+          kyberFeeHandler1.address,
+          kyberFeeHandler2.address,
+          kyberFeeHandler3.address,
+        ],
+        [ZERO_ADDRESS, rewardTokenA.address, rewardTokenB.address],
         {
           from: poolMasterOwner,
         }
       );
     });
 
-    it('should return 0 if PoolMaster has calledRewardMaster', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedReward = await kyberPoolMaster.claimedPoolReward(1);
+    it('should return 0 if PoolMaster has claimedPoolReward', async () => {
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler1.address);
+      const claimedReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
       expect(claimedReward).to.equal(true);
 
-      const unclaimed = await kyberPoolMaster.getUnclaimedRewards(1);
+      const unclaimed = await kyberPoolMaster.getUnclaimedRewards(
+        1,
+        kyberFeeHandler1.address
+      );
       expect(unclaimed.toString()).to.equal('0');
     });
 
     it("should return 0 if staker's reward percentage in precision for the epoch is 0", async () => {
-      const claimedReward = await kyberPoolMaster.claimedPoolReward(2);
+      const claimedReward = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler1.address
+      );
       expect(claimedReward).to.equal(false);
 
       await kyberDao.setStakerRewardPercentage(kyberPoolMaster.address, 2, 0);
@@ -86,12 +123,18 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       );
       expect(stakerReward.toString()).to.equal('0');
 
-      const unclaimed = await kyberPoolMaster.getUnclaimedRewards(2);
+      const unclaimed = await kyberPoolMaster.getUnclaimedRewards(
+        2,
+        kyberFeeHandler1.address
+      );
       expect(unclaimed.toString()).to.equal('0');
     });
 
     it('should return 0 if total reward for the epoch is 0', async () => {
-      const claimedReward = await kyberPoolMaster.claimedPoolReward(2);
+      const claimedReward = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler1.address
+      );
       expect(claimedReward).to.equal(false);
 
       await kyberDao.setStakerRewardPercentage(kyberPoolMaster.address, 2, 10);
@@ -101,16 +144,22 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       );
       expect(stakerReward.toString()).to.equal('10');
 
-      await kyberFeeHandler.setRewardsPerEpoch(2, 0);
-      const rewardPerEpoch = await kyberFeeHandler.rewardsPerEpoch(2);
+      await kyberFeeHandler1.setRewardsPerEpoch(2, 0);
+      const rewardPerEpoch = await kyberFeeHandler1.rewardsPerEpoch(2);
       expect(rewardPerEpoch.toString()).to.equal('0');
 
-      const unclaimed = await kyberPoolMaster.getUnclaimedRewards(2);
+      const unclaimed = await kyberPoolMaster.getUnclaimedRewards(
+        2,
+        kyberFeeHandler1.address
+      );
       expect(unclaimed.toString()).to.equal('0');
     });
 
     it('should return unclaimed reward amount', async () => {
-      const claimedReward = await kyberPoolMaster.claimedPoolReward(2);
+      const claimedReward = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler1.address
+      );
       expect(claimedReward).to.equal(false);
 
       await kyberDao.setStakerRewardPercentage(
@@ -124,18 +173,71 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       );
       expect(stakerReward.toString()).to.equal('200000000000000000'); // 20%
 
-      await kyberFeeHandler.setRewardsPerEpoch(2, '3000000000000000000'); // 3ETH
-      const rewardPerEpoch = await kyberFeeHandler.rewardsPerEpoch(2);
+      await kyberFeeHandler1.setRewardsPerEpoch(2, '3000000000000000000'); // 3ETH
+      const rewardPerEpoch = await kyberFeeHandler1.rewardsPerEpoch(2);
       expect(rewardPerEpoch.toString()).to.equal('3000000000000000000');
 
-      const unclaimed = await kyberPoolMaster.getUnclaimedRewards(2);
+      const unclaimed = await kyberPoolMaster.getUnclaimedRewards(
+        2,
+        kyberFeeHandler1.address
+      );
       expect(unclaimed.toString()).to.equal('600000000000000000'); // 3ETH -> 20% = 0.6ETH
+    });
+
+    it('should return all epochs with at least one FeeHandler paying rewards', async () => {
+      await kyberDao.setCurrentEpochNumber(7);
+
+      await kyberDao.setStakerRewardPercentage(
+        kyberPoolMaster.address,
+        3,
+        '200000000000000000'
+      ); // 20%
+      await kyberDao.setStakerRewardPercentage(kyberPoolMaster.address, 4, '0'); // 20%
+      await kyberDao.setStakerRewardPercentage(
+        kyberPoolMaster.address,
+        5,
+        '200000000000000000'
+      ); // 20%
+      await kyberDao.setStakerRewardPercentage(
+        kyberPoolMaster.address,
+        6,
+        '200000000000000000'
+      ); // 20%
+      await kyberDao.setStakerRewardPercentage(
+        kyberPoolMaster.address,
+        7,
+        '200000000000000000'
+      ); // 20%
+
+      await kyberFeeHandler1.setRewardsPerEpoch(3, '3000000000000000000'); // 3ETH
+      await kyberFeeHandler2.setRewardsPerEpoch(3, '0'); // 0ETH
+      await kyberFeeHandler3.setRewardsPerEpoch(3, '0'); // 0ETH
+
+      await kyberFeeHandler1.setRewardsPerEpoch(4, '3000000000000000000'); // 0ETH
+      await kyberFeeHandler2.setRewardsPerEpoch(4, '3000000000000000000'); // 0ETH
+      await kyberFeeHandler3.setRewardsPerEpoch(4, '3000000000000000000'); // 0ETH
+
+      await kyberFeeHandler1.setRewardsPerEpoch(5, '0'); // 0ETH
+      await kyberFeeHandler2.setRewardsPerEpoch(5, '0'); // 0ETH
+      await kyberFeeHandler3.setRewardsPerEpoch(5, '0'); // 0ETH
+
+      await kyberFeeHandler1.setRewardsPerEpoch(6, '0'); // 0ETH
+      await kyberFeeHandler2.setRewardsPerEpoch(6, '3000000000000000000'); // 3ETH
+      await kyberFeeHandler3.setRewardsPerEpoch(6, '3000000000000000000'); // 3ETH
+
+      await kyberFeeHandler1.setRewardsPerEpoch(7, '0'); // 0ETH
+      await kyberFeeHandler2.setRewardsPerEpoch(7, '3000000000000000000'); // 3ETH
+      await kyberFeeHandler3.setRewardsPerEpoch(7, '0'); // 0ETH
+
+      const unclaimedEpochs = await kyberPoolMaster.getAllEpochWithUnclaimedRewards();
+      expect(JSON.stringify(unclaimedEpochs)).to.equal('["2","3","6","7"]'); // epoch 2 is set in previos tests
     });
   });
 
   const prepareEpochForClaim = async ({
     epoch,
     staker,
+    feeHandlers,
     stakerRewardPercentage,
     rewardsPerEpoch,
     stakerStake = '0',
@@ -146,7 +248,11 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       epoch,
       stakerRewardPercentage
     );
-    await kyberFeeHandler.setRewardsPerEpoch(epoch, rewardsPerEpoch);
+    await Promise.all(
+      feeHandlers.map((feeHandler, i) =>
+        feeHandler.setRewardsPerEpoch(epoch, rewardsPerEpoch[i])
+      )
+    );
     await kyberStaking.setStakerData(
       epoch,
       staker,
@@ -170,15 +276,43 @@ contract('KyberPoolMaster claiming', async (accounts) => {
 
       kyberDao = await KyberDao.new(NO_ZERO_ADDRESS, kyberStaking.address);
 
-      kyberFeeHandler = await KyberFeeHandlerWithClaimStakerReward.new(
+      kyberFeeHandler1 = await KyberFeeHandlerWithClaimStakerReward.new(
         kyberDao.address
+      );
+
+      rewardTokenA = await TestToken.new('Reward Token A', 'RTA', 18);
+      kyberFeeHandler2 = await KyberFeeHandlerWithClaimStakerRewardERC20.new(
+        kyberDao.address,
+        rewardTokenA.address
+      );
+
+      rewardTokenB = await TestToken.new('Reward Token B', 'RTB', 18);
+      kyberFeeHandler3 = await KyberFeeHandlerWithClaimStakerRewardERC20.new(
+        kyberDao.address,
+        rewardTokenB.address
+      );
+
+      kyberFeeHandler4 = await KyberFeeHandlerWithClaimStakerRewardERC20.new(
+        kyberDao.address,
+        rewardTokenB.address
       );
 
       kyberPoolMaster = await KyberPoolMaster.new(
         kyberDao.address,
-        kyberFeeHandler.address,
         2,
         100, // Denominated in 1e4 units - 100 = 1%
+        [
+          kyberFeeHandler1.address,
+          kyberFeeHandler2.address,
+          kyberFeeHandler3.address,
+          kyberFeeHandler4.address,
+        ],
+        [
+          ZERO_ADDRESS,
+          rewardTokenA.address,
+          rewardTokenB.address,
+          rewardTokenB.address,
+        ],
         {from: poolMasterOwner}
       );
 
@@ -187,42 +321,26 @@ contract('KyberPoolMaster claiming', async (accounts) => {
         {value: '10000000000000000000', from: bank}
       );
 
-      await kyberFeeHandler.send('10000000000000000000', {from: bank});
+      await kyberFeeHandler1.send('10000000000000000000', {from: bank});
+      await rewardTokenA.transfer(kyberFeeHandler2.address, mulPrecision(100));
+      await rewardTokenB.transfer(kyberFeeHandler3.address, mulPrecision(100));
+      await rewardTokenB.transfer(kyberFeeHandler4.address, mulPrecision(100));
     });
 
-    it('should only be able to receive ETH from KyberFeeHandler', async () => {
+    it('should only be able to receive ETH from KyberFeeHandler1', async () => {
       await expectRevert(
         kyberPoolMaster.send('10', {from: mike}),
-        'only accept ETH from Kyber'
+        'only accept ETH from a KyberFeeHandler'
       );
     });
 
-    it('should revert if epoch rewards has been already claimed', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedReward = await kyberPoolMaster.claimedPoolReward(1);
-      expect(claimedReward).to.equal(true);
-
-      await expectRevert(
-        kyberPoolMaster.claimRewardsMaster(1, {from: mike}),
-        'cRMaster: rewards already claimed'
-      );
-    });
-
-    it('should revert if no unclaimed reward for the epoch', async () => {
-      // this makes getUnclaimedRewards to return 0
-      await kyberDao.setStakerRewardPercentage(kyberPoolMaster.address, 1, 0);
-      await expectRevert(
-        kyberPoolMaster.claimRewardsMaster(1, {from: mike}),
-        'cRMaster: no rewards to claim'
-      );
-    });
-
-    it('should revert if poolMaster can receive its share', async () => {
+    it('should revert when rewards is in ETH and poolMaster can not receive its share', async () => {
       await prepareEpochForClaim({
         epoch: 1,
         staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler1],
         stakerRewardPercentage: '200000000000000000', // 20%
-        rewardsPerEpoch: '3000000000000000000', // 3ETH
+        rewardsPerEpoch: ['3000000000000000000'], // 3ETH
       });
 
       await kyberPoolMaster.transferOwnership(
@@ -233,7 +351,7 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       );
 
       await expectRevert(
-        kyberPoolMaster.claimRewardsMaster(1, {from: mike}),
+        kyberPoolMaster.claimRewardsMaster([1], {from: mike}),
         'cRMaste: poolMaster share transfer failed'
       );
     });
@@ -242,8 +360,9 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       await prepareEpochForClaim({
         epoch: 1,
         staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler1],
         stakerRewardPercentage: '200000000000000000', // 20%
-        rewardsPerEpoch: '3000000000000000000', // 3ETH,
+        rewardsPerEpoch: ['3000000000000000000'], // 3ETH,
       });
 
       await kyberPoolMaster.transferOwnership(
@@ -254,7 +373,7 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       );
 
       await expectRevert(
-        kyberPoolMaster.claimRewardsMaster(1, {from: mike}),
+        kyberPoolMaster.claimRewardsMaster([1], {from: mike}),
         'cRMaste: poolMaster share transfer failed'
       );
 
@@ -262,7 +381,9 @@ contract('KyberPoolMaster claiming', async (accounts) => {
         poolMasterOwner
       );
 
-      const receipt = await kyberPoolMaster.claimRewardsMaster(1, {from: mike});
+      const receipt = await kyberPoolMaster.claimRewardsMaster([1], {
+        from: mike,
+      });
       expectEvent(receipt, 'MasterClaimReward');
     });
 
@@ -280,40 +401,175 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       await prepareEpochForClaim({
         epoch: 1,
         staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler1, kyberFeeHandler2],
         stakerRewardPercentage: '200000000000000000', // 20%
-        rewardsPerEpoch: rewardsPerEpoch, // 3ETH,
+        rewardsPerEpoch: [rewardsPerEpoch, rewardsPerEpoch], // 3ETH,
       });
 
-      const poolMasterOwnerBalance = await balance.current(poolMasterOwner);
+      await prepareEpochForClaim({
+        epoch: 2,
+        staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler3, kyberFeeHandler4],
+        stakerRewardPercentage: '200000000000000000', // 20%
+        rewardsPerEpoch: [rewardsPerEpoch, rewardsPerEpoch], // 3ETH,
+      });
 
-      const receipt = await kyberPoolMaster.claimRewardsMaster(1, {from: mike});
+      const poolMasterOwnerETHBalance = await balance.current(poolMasterOwner);
+      const poolMasterOwnerRTABalance = await rewardTokenA.balanceOf(
+        poolMasterOwner
+      );
+      const poolMasterOwnerRTBBalance = await rewardTokenB.balanceOf(
+        poolMasterOwner
+      );
+
+      const receipt = await kyberPoolMaster.claimRewardsMaster([1, 2], {
+        from: mike,
+      });
+
       expectEvent(receipt, 'MasterClaimReward', {
         epoch: '1',
+        feeHandler: kyberFeeHandler1.address,
         poolMaster: poolMasterOwner,
+        rewardToken: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
         totalRewards: unclaimReward,
         feeApplied: '100',
         feeAmount: feeAmount,
         poolMasterShare: '0',
       });
 
-      const poolMasterOwnerBalanceAfter = await balance.current(
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '1',
+        feeHandler: kyberFeeHandler2.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenA.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: '0',
+      });
+
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '2',
+        feeHandler: kyberFeeHandler3.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenB.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: '0',
+      });
+
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '2',
+        feeHandler: kyberFeeHandler4.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenB.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: '0',
+      });
+
+      const poolMasterOwnerETHBalanceAfter = await balance.current(
         poolMasterOwner
       );
-      const expectedBalance = poolMasterOwnerBalance.add(new BN(feeAmount));
-      expect(poolMasterOwnerBalanceAfter.toString()).to.equal(
-        expectedBalance.toString()
+      const poolMasterOwnerRTABalanceAfter = await rewardTokenA.balanceOf(
+        poolMasterOwner
+      );
+      const poolMasterOwnerRTBBalanceAfter = await rewardTokenB.balanceOf(
+        poolMasterOwner
       );
 
-      const poolMembersShare = await kyberPoolMaster.memberRewards(1);
-      const expectedPoolMembersShare = new BN(unclaimReward).sub(
+      const expectedETHBalance = poolMasterOwnerETHBalance.add(
         new BN(feeAmount)
       );
-      expect(poolMembersShare.totalRewards.toString()).to.equal(
-        expectedPoolMembersShare.toString()
+      expect(poolMasterOwnerETHBalanceAfter.toString()).to.equal(
+        expectedETHBalance.toString()
       );
 
-      const claimedPoolReward = await kyberPoolMaster.claimedPoolReward(1);
-      expect(claimedPoolReward).to.equal(true);
+      const expectedRTABalance = poolMasterOwnerRTABalance.add(
+        new BN(feeAmount)
+      );
+      expect(poolMasterOwnerRTABalanceAfter.toString()).to.equal(
+        expectedRTABalance.toString()
+      );
+
+      const expectedRTBBalance = poolMasterOwnerRTBBalance.add(
+        new BN(feeAmount).mul(new BN(2))
+      );
+      expect(poolMasterOwnerRTBBalanceAfter.toString()).to.equal(
+        expectedRTBBalance.toString()
+      );
+
+      //
+
+      const poolMembersETHShare = await kyberPoolMaster.memberRewards(
+        1,
+        kyberFeeHandler1.address
+      );
+      const expectedPoolETHMembersShare = new BN(unclaimReward).sub(
+        new BN(feeAmount)
+      );
+      expect(poolMembersETHShare.totalRewards.toString()).to.equal(
+        expectedPoolETHMembersShare.toString()
+      );
+
+      const claimedPoolETHReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
+      expect(claimedPoolETHReward).to.equal(true);
+
+      const poolMembersRTAShare = await kyberPoolMaster.memberRewards(
+        1,
+        kyberFeeHandler2.address
+      );
+      const expectedPoolRTAMembersShare = new BN(unclaimReward).sub(
+        new BN(feeAmount)
+      );
+      expect(poolMembersRTAShare.totalRewards.toString()).to.equal(
+        expectedPoolRTAMembersShare.toString()
+      );
+
+      const claimedPoolRTAReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler2.address
+      );
+      expect(claimedPoolRTAReward).to.equal(true);
+
+      //
+      const poolMembersRTBShare3 = await kyberPoolMaster.memberRewards(
+        2,
+        kyberFeeHandler3.address
+      );
+      const poolMembersRTBShare4 = await kyberPoolMaster.memberRewards(
+        2,
+        kyberFeeHandler4.address
+      );
+      const expectedPoolRTBMembersShare3 = new BN(unclaimReward).sub(
+        new BN(feeAmount)
+      );
+      const expectedPoolRTBMembersShare4 = new BN(unclaimReward).sub(
+        new BN(feeAmount)
+      );
+      expect(poolMembersRTBShare3.totalRewards.toString()).to.equal(
+        expectedPoolRTBMembersShare3.toString()
+      );
+      expect(poolMembersRTBShare4.totalRewards.toString()).to.equal(
+        expectedPoolRTBMembersShare4.toString()
+      );
+
+      const claimedPoolRTBReward3 = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler3.address
+      );
+      expect(claimedPoolRTBReward3).to.equal(true);
+
+      const claimedPoolRTBReward4 = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler4.address
+      );
+      expect(claimedPoolRTBReward4).to.equal(true);
     });
 
     it('should transfer fee + share to poolMaster if it has stake for the epoch', async () => {
@@ -333,44 +589,177 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       await prepareEpochForClaim({
         epoch: 1,
         staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler1, kyberFeeHandler2],
         stakerRewardPercentage: '200000000000000000', // 20%
-        rewardsPerEpoch: rewardsPerEpoch, // 3ETH,
+        rewardsPerEpoch: [rewardsPerEpoch, rewardsPerEpoch], // 3ETH,
         stakerStake: '1',
         delegatedStake: '1',
       });
 
-      const poolMasterOwnerBalance = await balance.current(poolMasterOwner);
+      await prepareEpochForClaim({
+        epoch: 2,
+        staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler3, kyberFeeHandler4],
+        stakerRewardPercentage: '200000000000000000', // 20%
+        rewardsPerEpoch: [rewardsPerEpoch, rewardsPerEpoch], // 3ETH,
+        stakerStake: '1',
+        delegatedStake: '1',
+      });
 
-      const receipt = await kyberPoolMaster.claimRewardsMaster(1, {from: mike});
+      const poolMasterOwnerETHBalance = await balance.current(poolMasterOwner);
+      const poolMasterOwnerRTABalance = await rewardTokenA.balanceOf(
+        poolMasterOwner
+      );
+      const poolMasterOwnerRTBBalance = await rewardTokenB.balanceOf(
+        poolMasterOwner
+      );
+
+      const receipt = await kyberPoolMaster.claimRewardsMaster([1, 2], {
+        from: mike,
+      });
+
       expectEvent(receipt, 'MasterClaimReward', {
         epoch: '1',
+        feeHandler: kyberFeeHandler1.address,
         poolMaster: poolMasterOwner,
+        rewardToken: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
         totalRewards: unclaimReward,
         feeApplied: '100',
         feeAmount: feeAmount,
         poolMasterShare: share,
       });
 
-      const poolMasterOwnerBalanceAfter = await balance.current(
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '1',
+        feeHandler: kyberFeeHandler2.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenA.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: share,
+      });
+
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '2',
+        feeHandler: kyberFeeHandler3.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenB.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: share,
+      });
+
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '2',
+        feeHandler: kyberFeeHandler4.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenB.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: share,
+      });
+
+      const poolMasterOwnerETHBalanceAfter = await balance.current(
         poolMasterOwner
       );
-      const expectedBalance = poolMasterOwnerBalance
+      const expectedETHBalance = poolMasterOwnerETHBalance
         .add(new BN(feeAmount))
         .add(new BN(share));
-      expect(poolMasterOwnerBalanceAfter.toString()).to.equal(
-        expectedBalance.toString()
+      expect(poolMasterOwnerETHBalanceAfter.toString()).to.equal(
+        expectedETHBalance.toString()
       );
 
-      const poolMembersShare = await kyberPoolMaster.memberRewards(1);
-      const expectedPoolMembersShare = new BN(unclaimReward)
+      const poolMembersETHShare = await kyberPoolMaster.memberRewards(
+        1,
+        kyberFeeHandler1.address
+      );
+      const expectedPoolMembersETHShare = new BN(unclaimReward)
         .sub(new BN(feeAmount))
         .div(new BN(2));
-      expect(poolMembersShare.totalRewards.toString()).to.equal(
-        expectedPoolMembersShare.toString()
+      expect(poolMembersETHShare.totalRewards.toString()).to.equal(
+        expectedPoolMembersETHShare.toString()
       );
 
-      const claimedPoolReward = await kyberPoolMaster.claimedPoolReward(1);
-      expect(claimedPoolReward).to.equal(true);
+      const claimedPoolETHReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
+      expect(claimedPoolETHReward).to.equal(true);
+
+      const poolMasterOwnerRTABalanceAfter = await rewardTokenA.balanceOf(
+        poolMasterOwner
+      );
+      const expectedRTABalance = poolMasterOwnerRTABalance
+        .add(new BN(feeAmount))
+        .add(new BN(share));
+      expect(poolMasterOwnerRTABalanceAfter.toString()).to.equal(
+        expectedRTABalance.toString()
+      );
+
+      const poolMembersRTAShare = await kyberPoolMaster.memberRewards(
+        1,
+        kyberFeeHandler2.address
+      );
+      const expectedPoolMembersRTAShare = new BN(unclaimReward)
+        .sub(new BN(feeAmount))
+        .div(new BN(2));
+      expect(poolMembersRTAShare.totalRewards.toString()).to.equal(
+        expectedPoolMembersRTAShare.toString()
+      );
+
+      const claimedPoolRTAReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler2.address
+      );
+      expect(claimedPoolRTAReward).to.equal(true);
+
+      //
+      const poolMasterOwnerRTBBalanceAfter = await rewardTokenB.balanceOf(
+        poolMasterOwner
+      );
+      const expectedRTBBalance = poolMasterOwnerRTBBalance
+        .add(new BN(feeAmount).mul(new BN(2)))
+        .add(new BN(share).mul(new BN(2)));
+      expect(poolMasterOwnerRTBBalanceAfter.toString()).to.equal(
+        expectedRTBBalance.toString()
+      );
+
+      const poolMembersRTBShare3 = await kyberPoolMaster.memberRewards(
+        2,
+        kyberFeeHandler3.address
+      );
+      const expectedPoolMembersRTBShare3 = new BN(unclaimReward)
+        .sub(new BN(feeAmount))
+        .div(new BN(2));
+      expect(poolMembersRTBShare3.totalRewards.toString()).to.equal(
+        expectedPoolMembersRTBShare3.toString()
+      );
+
+      const claimedPoolRTBReward3 = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler3.address
+      );
+      expect(claimedPoolRTBReward3).to.equal(true);
+      //
+      const poolMembersRTBShare4 = await kyberPoolMaster.memberRewards(
+        2,
+        kyberFeeHandler4.address
+      );
+      const expectedPoolMembersRTBShare4 = new BN(unclaimReward)
+        .sub(new BN(feeAmount))
+        .div(new BN(2));
+      expect(poolMembersRTBShare4.totalRewards.toString()).to.equal(
+        expectedPoolMembersRTBShare4.toString()
+      );
+
+      const claimedPoolRTBReward4 = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler4.address
+      );
+      expect(claimedPoolRTBReward4).to.equal(true);
     });
 
     it('poolMaster should receive all the reward if no one has delegated its stake to him', async () => {
@@ -389,18 +778,39 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       await prepareEpochForClaim({
         epoch: 1,
         staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler1, kyberFeeHandler2],
         stakerRewardPercentage: '200000000000000000', // 20%
-        rewardsPerEpoch: rewardsPerEpoch, // 3ETH,
+        rewardsPerEpoch: [rewardsPerEpoch, rewardsPerEpoch], // 3ETH,
         stakerStake: '1',
         delegatedStake: '0',
       });
 
-      const poolMasterOwnerBalance = await balance.current(poolMasterOwner);
+      await prepareEpochForClaim({
+        epoch: 2,
+        staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler3, kyberFeeHandler4],
+        stakerRewardPercentage: '200000000000000000', // 20%
+        rewardsPerEpoch: [rewardsPerEpoch, rewardsPerEpoch], // 3ETH,
+        stakerStake: '1',
+        delegatedStake: '0',
+      });
 
-      const receipt = await kyberPoolMaster.claimRewardsMaster(1, {from: mike});
+      const poolMasterOwnerETHBalance = await balance.current(poolMasterOwner);
+      const poolMasterOwnerRTABalance = await rewardTokenA.balanceOf(
+        poolMasterOwner
+      );
+      const poolMasterOwnerRTBBalance = await rewardTokenB.balanceOf(
+        poolMasterOwner
+      );
+
+      const receipt = await kyberPoolMaster.claimRewardsMaster([1, 2], {
+        from: mike,
+      });
       expectEvent(receipt, 'MasterClaimReward', {
         epoch: '1',
+        feeHandler: kyberFeeHandler1.address,
         poolMaster: poolMasterOwner,
+        rewardToken: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
         totalRewards: unclaimReward,
         feeApplied: '100',
         feeAmount: feeAmount,
@@ -409,28 +819,135 @@ contract('KyberPoolMaster claiming', async (accounts) => {
           .toString(),
       });
 
-      const poolMasterOwnerBalanceAfter = await balance.current(
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '1',
+        feeHandler: kyberFeeHandler2.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenA.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: new BN(unclaimReward)
+          .sub(new BN(feeAmount))
+          .toString(),
+      });
+
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '2',
+        feeHandler: kyberFeeHandler3.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenB.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: new BN(unclaimReward)
+          .sub(new BN(feeAmount))
+          .toString(),
+      });
+
+      expectEvent(receipt, 'MasterClaimReward', {
+        epoch: '2',
+        feeHandler: kyberFeeHandler4.address,
+        poolMaster: poolMasterOwner,
+        rewardToken: rewardTokenB.address,
+        totalRewards: unclaimReward,
+        feeApplied: '100',
+        feeAmount: feeAmount,
+        poolMasterShare: new BN(unclaimReward)
+          .sub(new BN(feeAmount))
+          .toString(),
+      });
+
+      const poolMasterOwnerETHBalanceAfter = await balance.current(
         poolMasterOwner
       );
-      const expectedBalance = poolMasterOwnerBalance.add(new BN(unclaimReward));
-
-      expect(poolMasterOwnerBalanceAfter.toString()).to.equal(
-        expectedBalance.toString()
+      const expectedETHBalance = poolMasterOwnerETHBalance.add(
+        new BN(unclaimReward)
       );
 
-      const poolMembersShare = await kyberPoolMaster.memberRewards(1);
-      expect(poolMembersShare.totalRewards.toString()).to.equal('0');
+      expect(poolMasterOwnerETHBalanceAfter.toString()).to.equal(
+        expectedETHBalance.toString()
+      );
 
-      const claimedPoolReward = await kyberPoolMaster.claimedPoolReward(1);
-      expect(claimedPoolReward).to.equal(true);
+      const poolMembersETHShare = await kyberPoolMaster.memberRewards(
+        1,
+        kyberFeeHandler1.address
+      );
+      expect(poolMembersETHShare.totalRewards.toString()).to.equal('0');
+
+      const claimedPoolETHReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
+      expect(claimedPoolETHReward).to.equal(true);
+
+      const poolMasterOwnerRTABalanceAfter = await rewardTokenA.balanceOf(
+        poolMasterOwner
+      );
+      const expectedRTABalance = poolMasterOwnerRTABalance.add(
+        new BN(unclaimReward)
+      );
+
+      expect(poolMasterOwnerRTABalanceAfter.toString()).to.equal(
+        expectedRTABalance.toString()
+      );
+
+      const poolMembersRTAShare = await kyberPoolMaster.memberRewards(
+        1,
+        kyberFeeHandler2.address
+      );
+      expect(poolMembersRTAShare.totalRewards.toString()).to.equal('0');
+
+      const claimedPoolRTAReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler2.address
+      );
+      expect(claimedPoolRTAReward).to.equal(true);
+      //
+
+      const poolMasterOwnerRTBBalanceAfter = await rewardTokenB.balanceOf(
+        poolMasterOwner
+      );
+      const expectedRTBBalance = poolMasterOwnerRTBBalance.add(
+        new BN(unclaimReward.mul(new BN(2)))
+      );
+
+      expect(poolMasterOwnerRTBBalanceAfter.toString()).to.equal(
+        expectedRTBBalance.toString()
+      );
+
+      const poolMembersRTBShare3 = await kyberPoolMaster.memberRewards(
+        2,
+        kyberFeeHandler3.address
+      );
+      expect(poolMembersRTBShare3.totalRewards.toString()).to.equal('0');
+
+      const claimedPoolRTBReward3 = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler3.address
+      );
+      expect(claimedPoolRTBReward3).to.equal(true);
+      //
+      const poolMembersRTBShare4 = await kyberPoolMaster.memberRewards(
+        2,
+        kyberFeeHandler4.address
+      );
+      expect(poolMembersRTBShare4.totalRewards.toString()).to.equal('0');
+
+      const claimedPoolRTBReward4 = await kyberPoolMaster.claimedPoolReward(
+        2,
+        kyberFeeHandler4.address
+      );
+      expect(claimedPoolRTBReward4).to.equal(true);
     });
 
     it('should apply the fee used if it was pending', async () => {
       await prepareEpochForClaim({
         epoch: 2,
         staker: kyberPoolMaster.address,
+        feeHandlers: [kyberFeeHandler1, kyberFeeHandler2],
         stakerRewardPercentage: '200000000000000000', // 20%
-        rewardsPerEpoch: ether('3'), // 3ETH,
+        rewardsPerEpoch: [ether('3'), ether('3')], // 3ETH
         stakerStake: '1',
         delegatedStake: '1',
       });
@@ -441,13 +958,16 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       });
 
       await kyberDao.setCurrentEpochNumber(2);
-      const receipt = await kyberPoolMaster.claimRewardsMaster(2, {from: mike});
+      const receipt = await kyberPoolMaster.claimRewardsMaster([2], {
+        from: mike,
+      });
       expectEvent(receipt, 'NewFees', {
         fromEpoch: '2',
         feeRate: '200',
       });
     });
   });
+
   describe('#getUnclaimedRewardsMember', () => {
     beforeEach('running before each test', async () => {
       await reverter.revert();
@@ -465,9 +985,10 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       );
       kyberPoolMaster = await KyberPoolMaster.new(
         kyberDao.address,
-        kyberFeeHandler.address,
         2,
         100, // Denominated in 1e4 units - 100 = 1%
+        [kyberFeeHandler.address],
+        [ZERO_ADDRESS],
         {from: poolMasterOwner}
       );
 
@@ -480,44 +1001,61 @@ contract('KyberPoolMaster claiming', async (accounts) => {
     });
 
     it('should return 0 if PoolMaster has not called claimRewardMaster', async () => {
-      const claimedReward = await kyberPoolMaster.claimedPoolReward(1);
+      const claimedReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler.address
+      );
       expect(claimedReward).to.equal(false);
 
       const unclaimed = await kyberPoolMaster.getUnclaimedRewardsMember(
         mike,
-        1
+        1,
+        kyberFeeHandler.address
       );
 
       expect(unclaimed.toString()).to.equal('0');
     });
 
     it('should return 0 if PoolMember has previously claimed reward for the epoch', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(1);
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler.address);
+      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler.address
+      );
       expect(claimedRewardMaster).to.equal(true);
 
-      await kyberPoolMaster.setClaimedDelegateReward(1, mike);
+      await kyberPoolMaster.setClaimedDelegateReward(
+        1,
+        mike,
+        kyberFeeHandler.address
+      );
       const claimedRewardMember = await kyberPoolMaster.claimedDelegateReward(
         1,
-        mike
+        mike,
+        kyberFeeHandler.address
       );
       expect(claimedRewardMember).to.equal(true);
 
       const unclaimed = await kyberPoolMaster.getUnclaimedRewardsMember(
         mike,
-        1
+        1,
+        kyberFeeHandler.address
       );
       expect(unclaimed.toString()).to.equal('0');
     });
 
     it('should return 0 if PoolMember has not stake for the epoch', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(1);
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler.address);
+      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler.address
+      );
       expect(claimedRewardMaster).to.equal(true);
 
       const claimedRewardMember = await kyberPoolMaster.claimedDelegateReward(
         1,
-        mike
+        mike,
+        kyberFeeHandler.address
       );
       expect(claimedRewardMember).to.equal(false);
 
@@ -527,19 +1065,24 @@ contract('KyberPoolMaster claiming', async (accounts) => {
 
       const unclaimed = await kyberPoolMaster.getUnclaimedRewardsMember(
         mike,
-        1
+        1,
+        kyberFeeHandler.address
       );
       expect(unclaimed.toString()).to.equal('0');
     });
 
     it('should return 0 if PoolMember has not delegated it stake to this contract for the epoch', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(1);
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler.address);
+      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler.address
+      );
       expect(claimedRewardMaster).to.equal(true);
 
       const claimedRewardMember = await kyberPoolMaster.claimedDelegateReward(
         1,
-        mike
+        mike,
+        kyberFeeHandler.address
       );
       expect(claimedRewardMember).to.equal(false);
 
@@ -550,19 +1093,24 @@ contract('KyberPoolMaster claiming', async (accounts) => {
 
       const unclaimed = await kyberPoolMaster.getUnclaimedRewardsMember(
         mike,
-        1
+        1,
+        kyberFeeHandler.address
       );
       expect(unclaimed.toString()).to.equal('0');
     });
 
     it('should return unclaimed poolMember reward amount', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(1);
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler.address);
+      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler.address
+      );
       expect(claimedRewardMaster).to.equal(true);
 
       const claimedRewardMember = await kyberPoolMaster.claimedDelegateReward(
         1,
-        mike
+        mike,
+        kyberFeeHandler.address
       );
       expect(claimedRewardMember).to.equal(false);
 
@@ -571,15 +1119,40 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       expect(stakerRawData[0].toString()).to.equal('1');
       expect(stakerRawData[2]).to.equal(kyberPoolMaster.address);
 
-      await kyberPoolMaster.setMemberRewards(1, 10, 5);
+      await kyberPoolMaster.setMemberRewards(1, kyberFeeHandler.address, 10, 5);
 
       const unclaimed = await kyberPoolMaster.getUnclaimedRewardsMember(
         mike,
-        1
+        1,
+        kyberFeeHandler.address
       );
       expect(unclaimed.toString()).to.equal('2');
     });
+
+    it('should return all epochs with pending to claim rewards', async () => {
+      await kyberDao.setCurrentEpochNumber(7);
+
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler.address);
+      await kyberPoolMaster.setClaimedPoolReward(2, kyberFeeHandler.address);
+      await kyberPoolMaster.setClaimedPoolReward(3, kyberFeeHandler.address);
+      await kyberPoolMaster.setClaimedPoolReward(4, kyberFeeHandler.address);
+      await kyberPoolMaster.setClaimedPoolReward(5, kyberFeeHandler.address);
+
+      await kyberStaking.setStakerData(1, mike, 1, 0, kyberPoolMaster.address);
+      await kyberStaking.setStakerData(3, mike, 1, 0, kyberPoolMaster.address);
+      await kyberStaking.setStakerData(5, mike, 1, 0, kyberPoolMaster.address);
+
+      await kyberPoolMaster.setMemberRewards(1, kyberFeeHandler.address, 10, 5);
+      await kyberPoolMaster.setMemberRewards(3, kyberFeeHandler.address, 10, 5);
+      await kyberPoolMaster.setMemberRewards(5, kyberFeeHandler.address, 10, 5);
+
+      const unclaimedEpochs = await kyberPoolMaster.getAllEpochWithUnclaimedRewardsMember(
+        mike
+      );
+      expect(JSON.stringify(unclaimedEpochs)).to.equal('["1","3","5"]');
+    });
   });
+
   describe('#claimRewardMember', () => {
     beforeEach('running before each test', async () => {
       await reverter.revert();
@@ -594,54 +1167,139 @@ contract('KyberPoolMaster claiming', async (accounts) => {
 
       kyberStaking = await KyberStakingWithgetStakerDataForEpoch.new();
       kyberDao = await KyberDao.new(NO_ZERO_ADDRESS, kyberStaking.address);
-      kyberFeeHandler = await KyberFeeHandlerWithClaimStakerReward.new(
+
+      kyberFeeHandler1 = await KyberFeeHandlerWithClaimStakerReward.new(
         kyberDao.address
+      );
+
+      rewardTokenA = await TestToken.new('Reward Token A', 'RTA', 18);
+      kyberFeeHandler2 = await KyberFeeHandlerWithClaimStakerRewardERC20.new(
+        kyberDao.address,
+        rewardTokenA.address
+      );
+
+      rewardTokenB = await TestToken.new('Reward Token B', 'RTB', 18);
+      kyberFeeHandler3 = await KyberFeeHandlerWithClaimStakerRewardERC20.new(
+        kyberDao.address,
+        rewardTokenB.address
+      );
+
+      kyberFeeHandler4 = await KyberFeeHandlerWithClaimStakerRewardERC20.new(
+        kyberDao.address,
+        rewardTokenB.address
       );
       kyberPoolMaster = await KyberPoolMaster.new(
         kyberDao.address,
-        kyberFeeHandler.address,
         2,
         100, // Denominated in 1e4 units - 100 = 1%
+        [
+          kyberFeeHandler1.address,
+          kyberFeeHandler2.address,
+          kyberFeeHandler3.address,
+          kyberFeeHandler4.address,
+        ],
+        [
+          ZERO_ADDRESS,
+          rewardTokenA.address,
+          rewardTokenB.address,
+          rewardTokenB.address,
+        ],
         {from: poolMasterOwner, value: '1000000000000000000'}
       );
+
+      await rewardTokenA.transfer(kyberPoolMaster.address, mulPrecision(100));
+      await rewardTokenB.transfer(kyberPoolMaster.address, mulPrecision(100));
     });
 
-    it('should revert if poolMaster has not claimed the epoch rewards yet', async () => {
-      const claimedReward = await kyberPoolMaster.claimedPoolReward(1);
+    it('should transfer nothing if poolMaster has not claimed the epoch rewards yet', async () => {
+      const claimedReward = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
       expect(claimedReward).to.equal(false);
 
-      await expectRevert(
-        kyberPoolMaster.claimRewardMember(1, {from: mike}),
-        'cRMember: poolMaster has not claimed yet'
-      );
+      const mikeBalance = await balance.current(mike);
+
+      const txInfo = await kyberPoolMaster.claimRewardMember([1], {from: mike});
+      const tx = await web3.eth.getTransaction(txInfo.tx);
+      const gasUsed = new BN(txInfo.receipt.gasUsed);
+      const gasPrice = new BN(tx.gasPrice);
+
+      const mikeBalanceAfter = await balance.current(mike);
+      const expectedBalance = mikeBalance.sub(gasUsed.mul(gasPrice));
+      expect(mikeBalanceAfter.toString()).to.equal(expectedBalance.toString());
     });
 
-    it('should revert if poolMember has already claimed its share for the epoch', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(1);
+    it('should transfer nothing if poolMember has already claimed its share for the epoch', async () => {
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler1.address);
+      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
       expect(claimedRewardMaster).to.equal(true);
 
-      await kyberPoolMaster.setClaimedDelegateReward(1, mike);
+      await kyberPoolMaster.setMemberRewards(
+        1,
+        kyberFeeHandler1.address,
+        10,
+        5
+      );
+
+      await kyberPoolMaster.setClaimedDelegateReward(
+        1,
+        mike,
+        kyberFeeHandler1.address
+      );
+      const claimedDelegateReward = await kyberPoolMaster.claimedDelegateReward(
+        1,
+        mike,
+        kyberFeeHandler1.address
+      );
+      expect(claimedDelegateReward).to.equal(true);
+
+      await kyberPoolMaster.setClaimedDelegateReward(
+        1,
+        mike,
+        kyberFeeHandler1.address
+      );
       const claimedRewardMember = await kyberPoolMaster.claimedDelegateReward(
         1,
-        mike
+        mike,
+        kyberFeeHandler1.address
       );
       expect(claimedRewardMember).to.equal(true);
 
-      await expectRevert(
-        kyberPoolMaster.claimRewardMember(1, {from: mike}),
-        'cRMember: rewards already claimed'
-      );
+      const mikeBalance = await balance.current(mike);
+
+      const txInfo = await kyberPoolMaster.claimRewardMember([1], {from: mike});
+      const tx = await web3.eth.getTransaction(txInfo.tx);
+      const gasUsed = new BN(txInfo.receipt.gasUsed);
+      const gasPrice = new BN(tx.gasPrice);
+
+      const mikeBalanceAfter = await balance.current(mike);
+      const expectedBalance = mikeBalance.sub(gasUsed.mul(gasPrice));
+      expect(mikeBalanceAfter.toString()).to.equal(expectedBalance.toString());
     });
 
-    it('should revert if unclaimed reward member is 0 due to the poolMember has not stake for the epoch', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(1);
+    it('should transfer nothing if unclaimed reward member is 0 due to the poolMember has not stake for the epoch', async () => {
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler1.address);
+      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
       expect(claimedRewardMaster).to.equal(true);
+
+      await kyberPoolMaster.setMemberRewards(
+        1,
+        kyberFeeHandler1.address,
+        10,
+        5
+      );
 
       const claimedRewardMember = await kyberPoolMaster.claimedDelegateReward(
         1,
-        mike
+        mike,
+        kyberFeeHandler1.address
       );
       expect(claimedRewardMember).to.equal(false);
 
@@ -649,20 +1307,37 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       const stakerRawData = await kyberStaking.getStakerRawData(mike, 1);
       expect(stakerRawData[0].toString()).to.equal('0');
 
-      await expectRevert(
-        kyberPoolMaster.claimRewardMember(1, {from: mike}),
-        'cRMember: no rewards to claim'
-      );
+      const mikeBalance = await balance.current(mike);
+
+      const txInfo = await kyberPoolMaster.claimRewardMember([1], {from: mike});
+      const tx = await web3.eth.getTransaction(txInfo.tx);
+      const gasUsed = new BN(txInfo.receipt.gasUsed);
+      const gasPrice = new BN(tx.gasPrice);
+
+      const mikeBalanceAfter = await balance.current(mike);
+      const expectedBalance = mikeBalance.sub(gasUsed.mul(gasPrice));
+      expect(mikeBalanceAfter.toString()).to.equal(expectedBalance.toString());
     });
 
-    it('should revert if unclaimed reward member is 0 due to the poolMember has not delegated it stake to this contract for the epoch', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(1);
+    it('should transfer nothing if unclaimed reward member is 0 due to the poolMember has not delegated it stake to this contract for the epoch', async () => {
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler1.address);
+      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
       expect(claimedRewardMaster).to.equal(true);
+
+      await kyberPoolMaster.setMemberRewards(
+        1,
+        kyberFeeHandler1.address,
+        10,
+        5
+      );
 
       const claimedRewardMember = await kyberPoolMaster.claimedDelegateReward(
         1,
-        mike
+        mike,
+        kyberFeeHandler1.address
       );
       expect(claimedRewardMember).to.equal(false);
 
@@ -671,20 +1346,30 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       expect(stakerRawData[0].toString()).to.equal('1');
       expect(stakerRawData[2]).not.to.equal(kyberPoolMaster.address);
 
-      await expectRevert(
-        kyberPoolMaster.claimRewardMember(1, {from: mike}),
-        'cRMember: no rewards to claim'
-      );
+      const mikeBalance = await balance.current(mike);
+
+      const txInfo = await kyberPoolMaster.claimRewardMember([1], {from: mike});
+      const tx = await web3.eth.getTransaction(txInfo.tx);
+      const gasUsed = new BN(txInfo.receipt.gasUsed);
+      const gasPrice = new BN(tx.gasPrice);
+
+      const mikeBalanceAfter = await balance.current(mike);
+      const expectedBalance = mikeBalance.sub(gasUsed.mul(gasPrice));
+      expect(mikeBalanceAfter.toString()).to.equal(expectedBalance.toString());
     });
 
     it('poolMember should receive its share', async () => {
-      await kyberPoolMaster.setClaimedPoolReward(1);
-      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(1);
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler1.address);
+      const claimedRewardMaster = await kyberPoolMaster.claimedPoolReward(
+        1,
+        kyberFeeHandler1.address
+      );
       expect(claimedRewardMaster).to.equal(true);
 
       const claimedRewardMember = await kyberPoolMaster.claimedDelegateReward(
         1,
-        mike
+        mike,
+        kyberFeeHandler1.address
       );
       expect(claimedRewardMember).to.equal(false);
 
@@ -693,14 +1378,21 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       expect(stakerRawData[0].toString()).to.equal('1');
       expect(stakerRawData[2]).to.equal(kyberPoolMaster.address);
 
-      await kyberPoolMaster.setMemberRewards(1, 10, 5);
+      await kyberPoolMaster.setMemberRewards(
+        1,
+        kyberFeeHandler1.address,
+        10,
+        5
+      );
 
       const mikeBalance = await balance.current(mike);
 
-      const txInfo = await kyberPoolMaster.claimRewardMember(1, {from: mike});
+      const txInfo = await kyberPoolMaster.claimRewardMember([1], {from: mike});
       expectEvent(txInfo, 'MemberClaimReward', {
         epoch: '1',
         poolMember: mike,
+        feeHandler: kyberFeeHandler1.address,
+        rewardToken: ETH_TOKEN_ADDRESS,
         reward: '2',
       });
 
@@ -715,6 +1407,109 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       expect(mikeBalanceAfter.toString()).to.equal(expectedBalance.toString());
     });
 
+    it('poolMember should receive its share from multiple feeHandlers claiming in multiple epochs', async () => {
+      //epoch 1 - fh1 - fh2
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler1.address);
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler2.address);
+
+      await kyberStaking.setStakerData(1, mike, 1, 0, kyberPoolMaster.address);
+
+      await kyberPoolMaster.setMemberRewards(
+        1,
+        kyberFeeHandler1.address,
+        10,
+        5
+      );
+      await kyberPoolMaster.setMemberRewards(
+        1,
+        kyberFeeHandler2.address,
+        10,
+        5
+      );
+
+      // epoch 2 - fh3 - fh4
+      await kyberPoolMaster.setClaimedPoolReward(2, kyberFeeHandler3.address);
+      await kyberPoolMaster.setClaimedPoolReward(2, kyberFeeHandler4.address);
+
+      await kyberStaking.setStakerData(2, mike, 1, 0, kyberPoolMaster.address);
+
+      await kyberPoolMaster.setMemberRewards(
+        2,
+        kyberFeeHandler3.address,
+        10,
+        5
+      );
+      await kyberPoolMaster.setMemberRewards(
+        2,
+        kyberFeeHandler4.address,
+        10,
+        5
+      );
+
+      const mikeETHBalance = await balance.current(mike);
+      const mikeRTABalance = await rewardTokenA.balanceOf(mike);
+      const mikeRTBBalance = await rewardTokenB.balanceOf(mike);
+
+      const txInfo = await kyberPoolMaster.claimRewardMember([1, 2], {
+        from: mike,
+      });
+      expectEvent(txInfo, 'MemberClaimReward', {
+        epoch: '1',
+        poolMember: mike,
+        feeHandler: kyberFeeHandler1.address,
+        rewardToken: ETH_TOKEN_ADDRESS,
+        reward: '2',
+      });
+
+      expectEvent(txInfo, 'MemberClaimReward', {
+        epoch: '1',
+        poolMember: mike,
+        feeHandler: kyberFeeHandler2.address,
+        rewardToken: rewardTokenA.address,
+        reward: '2',
+      });
+
+      expectEvent(txInfo, 'MemberClaimReward', {
+        epoch: '2',
+        poolMember: mike,
+        feeHandler: kyberFeeHandler3.address,
+        rewardToken: rewardTokenB.address,
+        reward: '2',
+      });
+
+      expectEvent(txInfo, 'MemberClaimReward', {
+        epoch: '2',
+        poolMember: mike,
+        feeHandler: kyberFeeHandler4.address,
+        rewardToken: rewardTokenB.address,
+        reward: '2',
+      });
+
+      const tx = await web3.eth.getTransaction(txInfo.tx);
+      const gasUsed = new BN(txInfo.receipt.gasUsed);
+      const gasPrice = new BN(tx.gasPrice);
+
+      const mikeETHBalanceAfter = await balance.current(mike);
+      const expectedETHBalance = mikeETHBalance
+        .sub(gasUsed.mul(gasPrice))
+        .add(new BN('2'));
+      expect(mikeETHBalanceAfter.toString()).to.equal(
+        expectedETHBalance.toString()
+      );
+
+      const mikeRTABalanceAfter = await rewardTokenA.balanceOf(mike);
+      const expectedRTABalance = mikeRTABalance.add(new BN('2'));
+      expect(mikeRTABalanceAfter.toString()).to.equal(
+        expectedRTABalance.toString()
+      );
+
+      const mikeRTBBalanceAfter = await rewardTokenB.balanceOf(mike);
+      const expectedRTBBalance = mikeRTBBalance.add(new BN('4'));
+      expect(mikeRTBBalanceAfter.toString()).to.equal(
+        expectedRTBBalance.toString()
+      );
+    });
+
     it('should distribute pool members share and keep remainings from rounding', async () => {
       // totalStaked = 100
       // total reward to share among members = 9999
@@ -725,7 +1520,7 @@ contract('KyberPoolMaster claiming', async (accounts) => {
       // remainings = 3
       // values are in wei
 
-      await kyberPoolMaster.setClaimedPoolReward(1);
+      await kyberPoolMaster.setClaimedPoolReward(1, kyberFeeHandler1.address);
       await kyberStaking.setStakerData(1, mike, 35, 0, kyberPoolMaster.address);
       await kyberStaking.setStakerData(1, cris, 5, 0, kyberPoolMaster.address);
       await kyberStaking.setStakerData(
@@ -736,37 +1531,48 @@ contract('KyberPoolMaster claiming', async (accounts) => {
         kyberPoolMaster.address
       );
 
-      await kyberPoolMaster.setMemberRewards(1, 9999, 100);
+      await kyberPoolMaster.setMemberRewards(
+        1,
+        kyberFeeHandler1.address,
+        9999,
+        100
+      );
 
       const mikeBalance = await balance.current(mike);
       const crisBalance = await balance.current(cris);
       const paulaBalance = await balance.current(paula);
       const poolBalance = await balance.current(kyberPoolMaster.address);
 
-      const mikeTxInfo = await kyberPoolMaster.claimRewardMember(1, {
+      const mikeTxInfo = await kyberPoolMaster.claimRewardMember([1], {
         from: mike,
       });
       expectEvent(mikeTxInfo, 'MemberClaimReward', {
         epoch: '1',
         poolMember: mike,
+        feeHandler: kyberFeeHandler1.address,
+        rewardToken: ETH_TOKEN_ADDRESS,
         reward: '3499',
       });
 
-      const crisTxInfo = await kyberPoolMaster.claimRewardMember(1, {
+      const crisTxInfo = await kyberPoolMaster.claimRewardMember([1], {
         from: cris,
       });
       expectEvent(crisTxInfo, 'MemberClaimReward', {
         epoch: '1',
         poolMember: cris,
+        feeHandler: kyberFeeHandler1.address,
+        rewardToken: ETH_TOKEN_ADDRESS,
         reward: '499',
       });
 
-      const paulaTxInfo = await kyberPoolMaster.claimRewardMember(1, {
+      const paulaTxInfo = await kyberPoolMaster.claimRewardMember([1], {
         from: paula,
       });
       expectEvent(paulaTxInfo, 'MemberClaimReward', {
         epoch: '1',
         poolMember: paula,
+        feeHandler: kyberFeeHandler1.address,
+        rewardToken: ETH_TOKEN_ADDRESS,
         reward: '5999',
       });
 
@@ -818,3 +1624,7 @@ contract('KyberPoolMaster claiming', async (accounts) => {
     });
   });
 });
+
+function mulPrecision(value) {
+  return precisionUnits.mul(new BN(value));
+}
